Pull request #45: OIFS-615 Fix incorrect boundary layer index calculation for nucleation
Merge in ~NM6/openifs-48r1 from bugfix/OIFS-615-incorrect-boundary-layer-index-calculation-for-nucleation to knmi-m7

Squashed commit of the following:

commit 18bdc28a1eb7ba10bbec098ec4f5d520c6e57105
Author: tommibergman <tommi.bergman@iki.fi>
Date:   Tue Sep 2 17:37:55 2025 +0300

     int -> nint for index of pbl top level

commit 27ccfad875fdf9826bca82812fe0205f8e7f3c7f
Author: tommibergman <tommi.bergman@iki.fi>
Date:   Fri Aug 29 13:30:06 2025 +0300

    OIFS-615 PR review corrections

commit d30fd5df7e6bb4a90e8fe575b48e4b3fa617b2c5
Author: tommibergman <tommi.bergman@iki.fi>
Date:   Thu Aug 21 11:43:05 2025 +0300

    OIFS-615 clear out writouts/comments on pblh index calc.'

commit 70244f0ef831f4f1022fe21542c1ae38d22aaef8
Author: tommibergman <tommi.bergman@iki.fi>
Date:   Thu Aug 21 11:30:43 2025 +0300

    OIFS-615 boundary layer index calculation changed to reflect that the PGEOH is the bottom of the grid box.

+++ ifs-source/arpifs/m7/module/mo_ham_m7_nucl.F90
@@ -720,12 +720,12 @@ SUBROUTINE nucl_vehkamaeki(kproma, kbdim,   klev,        & ! ECHAM5 dimensions
     
     DO jl = 1, kproma
       ! Above the boundary layer:
-      DO jk = 1, int(ppbl(jl)) - 1
+      DO jk = 1, nint(ppbl(jl)) - 1
         pns(jl,jk)  = 0.0_dp
         ppfr(jl,jk) = 0.0_dp
       ENDDO
       ! In the boundary layer:
-      DO jk = int(ppbl(jl)), klev
+      DO jk = nint(ppbl(jl)), klev
         pns(jl,jk)  = 1.0_dp
         ppfr(jl,jk) = 1.0E-6_dp*ph2so4(jl,jk)*pforest(jl)
       ENDDO
@@ -792,12 +792,12 @@ SUBROUTINE nucl_vehkamaeki(kproma, kbdim,   klev,        & ! ECHAM5 dimensions
     
     DO jl = 1, kproma
       ! Above the boundary layer:
-      DO jk = 1, int(ppbl(jl)) - 1
+      DO jk = 1, nint(ppbl(jl)) - 1
         pns(jl,jk)  = 0.0_dp
         ppfr(jl,jk) = 0.0_dp
       ENDDO
       ! In the boundary layer:
-      DO jk = int(ppbl(jl)), klev
+      DO jk = nint(ppbl(jl)), klev
         pns(jl,jk)  = 2.0_dp
         ppfr(jl,jk) = 3.86E-13_dp*ph2so4(jl,jk)*ph2so4(jl,jk)*pforest(jl)
       ENDDO
